Python_aiAnalysis/requirements.txt
# Redis
redis>=5.0.0

# Socket.IO Client
python-socketio[client]>=5.11.0

# AI/ML (根據你的需求調整版本)
transformers>=4.35.0
torch>=2.0.0

# Database
psycopg2-binary>=2.9.9

# 其他可能需要的
requests>=2.31.0